DataStoreCryptography/requirements.txt
Flask
python-keystoneclient
python-swiftclient
pydes